move examples to folder /examples
