+++ app.go
@@ -5,7 +5,7 @@ import (
 	"github.com/Elyart-Network/NyaBot/internal/server"
 	// === Import plugins here ===
 	_ "github.com/Elyart-Network/NyaBot/plugin/FlarumGroup" // Help Plugin
-	// _ "github.com/Elyart-Network/NyaBot/plugin/Example"        // Example Plugin
+	// _ "github.com/Elyart-Network/NyaBot/examples/plugin"        // Example Plugin
 	// === Internal Plugins ===
 	// _ "github.com/Elyart-Network/NyaBot/internal/botcli" // Bot CLI
 	// _ "github.com/Elyart-Network/NyaBot/internal/lua"    // Lua Plugin
